로직 개선
