+++ lib/feature/map/PoliceMapScreen.dart
@@ -1,14 +1,14 @@
 import 'dart:async';
-import 'package:blueberry_flutter_template/providers/policeMap/PoliceStationProvider.dart';
-import 'package:blueberry_flutter_template/widgets/policeMap/SendMessageWidget.dart';
+import 'package:blueberry_flutter_template/feature/map/provider/LocationProvider.dart';
+import 'package:blueberry_flutter_template/feature/map/provider/PermissionProvider.dart';
+import 'package:blueberry_flutter_template/feature/map/provider/PoliceStationProvider.dart';
+import 'package:blueberry_flutter_template/feature/map/widget/GoogleMapWidget.dart';
+import 'package:blueberry_flutter_template/feature/map/widget/PermissionDeniedWidget.dart';
+import 'package:blueberry_flutter_template/feature/map/widget/PoliceStationListWidget.dart';
+import 'package:blueberry_flutter_template/feature/map/widget/SendMessageWidget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:blueberry_flutter_template/providers/policeMap/LocationProvider.dart';
-import 'package:blueberry_flutter_template/providers/policeMap/PermissionProvider.dart';
-import 'package:blueberry_flutter_template/widgets/policeMap/GoogleMapWidget.dart';
-import 'package:blueberry_flutter_template/widgets/policeMap/PoliceStationListWidget.dart';
-import 'package:blueberry_flutter_template/widgets/policeMap/PermissionDeniedWidget.dart';
 
 class PoliceMapScreen extends ConsumerStatefulWidget {
   const PoliceMapScreen({super.key});
@@ -39,48 +39,46 @@ class _PoliceMapScreenState extends ConsumerState<PoliceMapScreen> {
         ref.watch(policeStationsProvider(locationState));
 
     return Scaffold(
-      body: permissionStatus == LocationPermissionStatus.initial
-          ? Center(child: CircularProgressIndicator())
-          : permissionStatus == LocationPermissionStatus.granted
-              ? SafeArea(
-                  child: Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Column(
-                      children: [
-                        Expanded(
-                          child: policeStationsAsyncValue.when(
-                            data: (policeStations) => GoogleMapWidget(
-                              googleMapControllerCompleter:
-                                  _googleMapControllerCompleter,
-                              locationState: locationState,
-                              policeStationsAsyncValue:
-                                  policeStationsAsyncValue,
-                            ),
-                            loading: () =>
-                                Center(child: CircularProgressIndicator()),
-                            error: (error, stack) =>
-                                Center(child: Text('Error: $error')),
-                          ),
-                        ),
-                        SizedBox(
-                          height: 8,
-                        ),
-                        Expanded(
-                          child: PoliceStationListWidget(
-                            googleMapControllerCompleter:
-                                _googleMapControllerCompleter,
-                            locationState: locationState,
-                            policeStationsAsyncValue: policeStationsAsyncValue,
-                          ),
-                        ),
-                        SendMessage(
-                          locationState: locationState,
-                        )
-                      ],
+      body: () {
+        if (permissionStatus == LocationPermissionStatus.initial) {
+          return Center(child: CircularProgressIndicator());
+        } else if (permissionStatus == LocationPermissionStatus.granted) {
+          return SafeArea(
+            child: Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: Column(
+                children: [
+                  Expanded(
+                    child: policeStationsAsyncValue.when(
+                      data: (policeStations) => GoogleMapWidget(
+                        googleMapControllerCompleter:
+                            _googleMapControllerCompleter,
+                        locationState: locationState,
+                        policeStationsAsyncValue: policeStationsAsyncValue,
+                      ),
+                      loading: () => Center(child: CircularProgressIndicator()),
+                      error: (error, stack) =>
+                          Center(child: Text('Error: $error')),
                     ),
                   ),
-                )
-              : PermissionDeniedWidget(permissionStatus: permissionStatus),
+                  SizedBox(height: 8),
+                  Expanded(
+                    child: PoliceStationListWidget(
+                      googleMapControllerCompleter:
+                          _googleMapControllerCompleter,
+                      locationState: locationState,
+                      policeStationsAsyncValue: policeStationsAsyncValue,
+                    ),
+                  ),
+                  SendMessage(locationState: locationState),
+                ],
+              ),
+            ),
+          );
+        } else {
+          return PermissionDeniedWidget(permissionStatus: permissionStatus);
+        }
+      }(),
     );
   }
 }
