+++ lib/feature/map/widget/GoogleMapWidget.dart
@@ -1,10 +1,11 @@
 import 'dart:async';
+import 'package:blueberry_flutter_template/feature/map/provider/LocationProvider.dart';
+import 'package:blueberry_flutter_template/feature/map/provider/PoliceStationProvider.dart';
+import 'package:blueberry_flutter_template/feature/map/provider/SeletedPlaceProvider.dart';
+import 'package:blueberry_flutter_template/model/GoogleMapPlace.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:blueberry_flutter_template/providers/policeMap/LocationProvider.dart';
-import 'package:blueberry_flutter_template/providers/policeMap/PoliceStationProvider.dart';
-import 'package:blueberry_flutter_template/providers/policeMap/SeletedPlaceProvider.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class GoogleMapWidget extends ConsumerWidget {
