+++ lib/feature/map/widget/PoliceStationListWidget.dart
@@ -1,9 +1,10 @@
 import 'dart:async';
-import 'package:blueberry_flutter_template/providers/policeMap/SeletedPlaceProvider.dart';
+import 'package:blueberry_flutter_template/feature/map/provider/PoliceStationProvider.dart';
+import 'package:blueberry_flutter_template/feature/map/provider/SeletedPlaceProvider.dart';
+import 'package:blueberry_flutter_template/model/GoogleMapPlace.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:blueberry_flutter_template/providers/policeMap/PoliceStationProvider.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class PoliceStationListWidget extends ConsumerWidget {
